Add package removal functionality
A major enhancement has been made to the project. This commit introduces functionality for removing a package from the project. The included changes allow the user to destroy infrastructure resources, remove environment variables associated with the package, and update the project state. Infrastructure files have been updated to cater for the new 'destroy' method.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>