Add support for managing project environment variables
This update introduces methods to add and remove environment variables at the project level. It also adds the ability to resolve environment variables and run commands with resolved variables. The "@esm2cjs/execa" dependency has been added in this commit. Methods to initialize Hereya in a project directory, add and remove a package from the project, and install necessary resources for hereya operations in an infrastructure have also been introduced. Several related tests and documentation updates are included too.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>