Implement package addition to workspace
Extended the local backend by adding functionalities to add, get and save workspace environment data. This included the creation of new methods and enhancement of existing methods. New functions were also implemented to provision and destroy packages related to a workspace in the infrastructure. New command functionalities added for workspace as well.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>