Fix remove command tests
The tests for the 'remove' command have been updated to include setup for testing the home directory. This includes stubbing the 'homedir' method to return a mock home directory and adding setup steps for creating necessary files and directories in the mock home directory. The teardown steps were updated to delete these additional files and directories after the test.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>