Add deploy functionality and related tests
A new 'deploy' functionality was implemented in the infrastructure code, including the provision for skipping deployment. Accommodations for deployment were added to the 'add', 'remove' and 'deploy' command files. Additionally, a test file for 'deploy' was created. The package.json file was updated to use npx for command generation and the configuration files were updated to include a deploy flag.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>