Implement secret markings for environment variables
This commit introduces the ability to mark certain environment variables as secrets. A "markSecret" option was added to the GetProjectEnvInput type and resolveEnvValues function, which if set to true, will prepend "secret://" to the value of any variable marked as secret. Further, support for path resolution has been added to projectRootDir, and local infrastructure now supports deploy and undeploy methods.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>