feat: Add support for deployment companion packages
Extended the system to handle a new attribute 'onDeploy' in the package metadata which indicates a deployment companion package. This is included in all necessary commands and their respective tests. Also added support for overriding infra while maintaining reference to the original infra. The commit involves changes in handling metadata, provisioning/destroying packages and updates in command flags.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>