fix: Store and delete parameters securely in AWS Parameter Store
The AWS setup has been updated to create secure parameters for each build operation. The parameters are stored in the AWS Parameter Store as 'SecureString', used in the build operation, and deleted immediately after. This greatly improves the security of our code builds by not exposing the parameters in plaintext.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>