Add Terraform binary downloading support
In the Terraform class, methods have been modified to use a downloaded Terraform binary instead of relying on a globally installed binary. Added a `downloadTerraform` method to handle the downloading of the appropriate binary based on the operating system and architecture. Also added the `unzip-stream` package and its types to handle decompression of the downloaded zip file.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>