fix: Update parameter formatting in AWS script
This commit adjusts the way parameters are formatted in the AWS script. Previously, the parameters were simply joined with a comma. Now, the value of the parameters is transformed into a JSON string before being joined, ensuring that complex objects or arrays can be properly represented and parsed when used later on.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>