fix: Refactor parameter and environment variables handling
This change includes refactoring on the handling of both parameters and environment variables. Now, if the values are of object type, they will be stringified. Otherwise, the original value is used. This applies to code in both `aws.ts` and `terraform.ts` files, making sure to serialize only the values that are objects.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>