fix: Implement Base64 encoding for object values
The changes include replacing JSON.stringify() with objectToBase64() for better handling of object values. In addition, a new function 'tryBase64ToJSONString' was defined in 'object-utils' and applied in 'aws.ts' and 'exec/index.ts' to safely decode Base64 and return the original input if the process fails. The test for this new function was also written.

Signed-off-by: Jonatan Sombié <jonatan.sombie@gmail.com>